Move lsp tests into integration folder and the other tests in integration_tests.rs
