+++ cli/tests/integration/lsp_tests.rs
